chore(build): replace use of repository archive during build-prod.
Downloading an archive of the repository is unconventional.
Typically, the sources for the build are sourced from the docker
context (`COPY .` - as done in `build-dev`). This was done (in #2002)
to protect production builds from incorporating source from
unclean working copies.

The most immediate downside to this is that this circumvents
the directives of `.dockerignore` (a recipe for expected behavior).
Otherwise, it just adds indirection.

The better approach is to only perform these builds in a clean-room
environment. For example, using build scripts that download and
unpack the archive in `/tmp` and building in that working directory
or, better yet, using CI/CD to build in a trusted environment.

Today, building (#2118) and deployment (#2169) are performed by
Github Actions. The concern about unclean working copies doesn't
have the force that it once did.