build(deps): update django-rest-framework for CSP
`django-rest-framework` has recieved several patches for
CSP compatibility, but unfortunately these are not included
in the latest release (3.14.0) and a new release cannot be
expected for a while.

As such, we must replace the semantically versioned PyPI
reference with a reference to the latest commit on the master
branch of the package's Git Repository (do not use `master`
since it could change in unexpected and incompatible ways).

Fixes: #3020