feat(build): eliminate latest-web-dev image
`docker-compose` and `docker compose` are substantially
different. The former corresponds with "Compose V1", which
is deprecated, and the later is a rewrite that,
significantly, uses the same build system as the other
`docker/*` actions and, thereby, the same cache.

Once done, it became apparent that the development
images are unnecessary.
- Pushing to `localhost:5000` in workflows isn't
  needed, because images produced by earlier steps
  are now visible to `docker compose`.
- With `localhost:5000` removed, `DJANGO_DOCKER_IMAGE`
  and `CELERY_DOCKER_IMAGE` environment variables are
  no longer used.
- With no consumption of development images outside
  of the local environment, there's no reason to
  publish these images.