chore(build): remove nonoperative `VENV_PATH`
Installing dependencies in a virtual environment has
merit, but declaring an environmental variable does
not accomplish this. Packages are installed in the
system root, as `POETRY_VIRTUALENVS_CREATE=false`
dictates.

```
% docker run -it freelawproject/courtlistener:latest-web-dev find /opt/pysetup
/opt/pysetup
/opt/pysetup/pyproject.toml
/opt/pysetup/poetry.lock
```

Removing the environment variable is the conservative
approach - it acknowledges the current behavior. The
alternative, which is best saved for a later commit is
to create the missing virtual environment:
```
python -m venv $VENV_PATH
```